requirements.txt
Django==5.2.8
psycopg[binary]>=3.2
django-csp>=4.0
djangorestframework==3.15.2
